Add files via upload
Run the Run.m file to start the refocus process. Parameters such as matrix size and image resolution can be set on the same file. The output can also be changed to other file types. The input must be a matrix with minor displacements along the X and Y axis